requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    ./tools/scripts/create_requirements.sh
#
attrs==21.4.0
    # via jsonschema
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via typer
future-fstrings==1.2.0
    # via pyecore
idna==3.3
    # via requests
importlib-resources==5.4.0
    # via jsonschema
jsonschema==4.3.2
    # via securicad-model
lxml==4.7.1
    # via pyecore
ordered-set==4.0.2
    # via pyecore
pyecore==0.12.2
    # via
    #   -r requirements.in
    #   securicad-model
pyrsistent==0.18.0
    # via jsonschema
requests==2.26.0
    # via securicad-enterprise
restrictedpython==5.2
    # via pyecore
securicad-enterprise==0.3.2
    # via -r requirements.in
securicad-model==1.0.0
    # via
    #   -r requirements.in
    #   securicad-enterprise
typer==0.4.0
    # via -r requirements.in
urllib3==1.26.7
    # via requests
wheel==0.37.1
    # via -r requirements.in
zipp==3.6.0
    # via importlib-resources
